requirements.txt
discord.py==2.5.2
dotenv==0.9.9
asyncio==3.4.3
requests==2.32.4